Skip sending stub mails for a number of PR events
The volume of mails from PRs on the *-commits mailing lists is
absolute massive, and many of them don't contain much actionable
information either.

E.g. when creating a new PR, this generates one mail with the
"labeled" event for each label that was added.

Reduce the number of emails sent; skip the stub mails for events
which probably aren't of relevance for the firehose style *-commits
mailing lists.

(Ideally, many consecutive mails would be coalesced, but I presume
that this isn't really trivial to do with how Github's notifications
work.)