Skip a trailing colon after the URL to a PR
When the plain (non-hyperlink) URL is followed by a colon (in
a plaintext email), the trailing colon is treated as part of the
url; at least Alpine for text based emails, and Apple Mail for
GUI clients, treat the colon as part of the URL here. This means
that following the URL fails unless manual effort is made to
remove the colon.