Ignore merge commits in PR Emailer
When merging main into a PR, the merge commits bring in changes to
files in all subprojects. Essentially this means that any PR that
merges main ends up sending mails to all mailing lists.

When iterating over the commits in the PR, skip the merge commits.
If the merge commits contains functional changes other than the
pure merge, the same changed files should be present in other
commits anyway, if the changes are due to conflicts.